Refactor the structure of `//config`
