+++ config/Config.java
@@ -65,15 +65,15 @@ public class Config {
     }
 
     public static class TraceSampling {
-        private Schema.SamplingFunction function;
+        private SamplingFunction function;
         private Integer arg;
 
         public Function<Integer, Boolean> getSamplingFunction() {
-            return Schema.SamplingFunction.applyArg(function, arg);
+            return SamplingFunction.applyArg(function, arg);
         }
 
         public void setFunction(String function) {
-            this.function = Schema.SamplingFunction.getByName(function);
+            this.function = SamplingFunction.getByName(function);
         }
 
         public Integer getArg() {
@@ -86,10 +86,10 @@ public class Config {
     }
 
     public static class Agent {
-        private Schema.AgentMode agentMode;
+        private AgentMode agentMode;
         private String name;
 
-        public static Agent ConstructAgentConfig(String name, Schema.AgentMode agentMode) {
+        public static Agent ConstructAgentConfig(String name, AgentMode agentMode) {
             // This method is needed because snakeyaml doesn't support declaring a constructor for this class
             Agent agent = new Agent();
             agent.setName(name);
@@ -108,21 +108,22 @@ public class Config {
         public void setMode(String mode) {
             switch (mode) {
                 case "trace":
-                    this.agentMode = Schema.AgentMode.TRACE;
+                    this.agentMode = AgentMode.TRACE;
                     break;
                 case "run":
-                    this.agentMode = Schema.AgentMode.RUN;
+                    this.agentMode = AgentMode.RUN;
                     break;
                 case "off":
-                    this.agentMode = Schema.AgentMode.OFF;
+                    this.agentMode = AgentMode.OFF;
                     break;
                 default:
                     throw new IllegalArgumentException(String.format("Unrecognised agent mode %s", mode));
             }
         }
 
-        public Schema.AgentMode getAgentMode() {
+        public AgentMode getAgentMode() {
             return agentMode;
         }
+
     }
 }
